requirements.txt
streamlit
google-generativeai
psycopg2-binary
python-dotenv
pandas
plotly
openpyxl
fpdf2
